Moving event emitters/observers to a separate directory (and outside of hooks)
