+++ packages/app-bridge/src/AppBridgeTheme.ts
@@ -65,7 +65,12 @@ export interface AppBridgeTheme extends AppBridgeBase {
 
     deleteStandardDocument(id: number): Promise<void>;
 
-    moveDocument(documentId: number, position: number, newDocumentGroupId?: number): Promise<Document>;
+    moveDocument(
+        documentId: number,
+        position: number,
+        newDocumentGroupId?: number,
+        oldGroupId?: number,
+    ): Promise<Document>;
 
     createDocumentGroup(documentGroup: DocumentGroupCreate): Promise<DocumentGroup>;
