+++ packages/app-bridge/src/types/index.ts
@@ -1,6 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './Asset';
+export * from './BlockSettings';
 export * from './BrandportalLink';
 export * from './BulkDownload';
 export * from './Color';
